+++ Sources/Wampproto/Messages/Message.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 public protocol Message: Sendable {
-    var type: Int64 { get }
+    var type: UInt64 { get }
     static func parse(message: [any Sendable]) throws -> Message
     func marshal() -> [any Sendable]
 }
@@ -9,5 +9,5 @@ public protocol Message: Sendable {
 public protocol BinaryPayload {
     var payloadIsBinary: Bool { get }
     var payload: Data? { get }
-    var payloadSerializer: Int { get }
+    var payloadSerializer: UInt64 { get }
 }
